// packages/connection-types-shared/src/concepts/connectionTypes/fields/types.ts
import type { ConnectionTypeDataField } from '../types';

export type FieldMode = 'preview' | 'default' | 'instance';

export type FieldProps<T extends ConnectionTypeDataField> = {
  id: string;
  field: T;
  'data-testid'?: string;
  mode?: FieldMode;
  onChange?: (value: T['properties']['defaultValue']) => void;
  value?: T['properties']['defaultValue'];
  onValidate?: (error: boolean | string, value?: T['properties']['defaultValue']) => void;
  error?: string | boolean;
  isDisabled?: boolean;
};