// packages/connection-types-shared/src/concepts/connectionTypes/types.ts
export enum ConnectionTypeFieldType {
  Boolean = 'boolean',
  Dropdown = 'dropdown',
  File = 'file',
  Hidden = 'hidden',
  Numeric = 'numeric',
  Section = 'section',
  ShortText = 'short-text',
  Text = 'text',
  URI = 'uri',
}
// ...
type Field<T extends ConnectionTypeFieldTypeUnion> = {
  type: T;
  name: string;
  description?: string;
};
// ...
export type SectionField = Field<ConnectionTypeFieldType.Section | 'section'>;

export type HiddenField = DataField<ConnectionTypeFieldType.Hidden | 'hidden'>;
export type ShortTextField = DataField<ConnectionTypeFieldType.ShortText | 'short-text'>;
export type TextField = DataField<ConnectionTypeFieldType.Text | 'text'>;
export type UriField = DataField<ConnectionTypeFieldType.URI | 'uri'>;
export type FileField = DataField<
  ConnectionTypeFieldType.File | 'file',
  string,
  {
    extensions?: string[];
    helperText?: string; // No UI to add this field, must do in yaml
  }
>;
export type BooleanField = DataField<
  ConnectionTypeFieldType.Boolean | 'boolean',
  boolean,
  {
    label?: string;
  }
>;
export type DropdownField = DataField<
  ConnectionTypeFieldType.Dropdown | 'dropdown',
  string[],
  {
    variant?: 'single' | 'multi';
    items?: { label: string; value: string }[];
  }
>;
export type NumericField = DataField<
  ConnectionTypeFieldType.Numeric | 'numeric',
  number,
  {
    unit?: string;
    min?: number;
    max?: number;
  }
>;

export type ConnectionTypeField =
  | BooleanField
  | DropdownField
  | FileField
  | HiddenField
  | NumericField
  | TextField
  | SectionField
  | ShortTextField
  | UriField;

export type ConnectionTypeDataField = Exclude<ConnectionTypeField, SectionField>;